Fix warnings when building Publish target
The Web.config transforms used to remove the configSource atribute
from the <connectionStrings> element even though it's not actually
set. You do need to strip it out if you start using configSource
as a way to specify locally configured connection strings, but it
was wrong to have that in the transform when people aren't using
it.