Fix race condition with image upload
The MediaUploader.js file reports progress and completion through a
single callback - you know you're done when you get a progress
notification telling you that all bytes have been uploaded. However,
there was a problem with that - it was reporting that all bytes had
been uploaded before it committed the blob. (When uploading directly
to Azure Blob Storage, although you can upload the entire blob at
once, that doesn't work well for video because there's a limit on
the length - 64MB, IIRC, so in practice you upload in chunks, and
then there's a 'commit' after the final chunk to tell Azure that
you've uploaded everything and it should make the data available.)

The problem was that this sometimes meant the server attempted to
build a thumbnail for an image that wasn't yet available - all the
data had been uploaded, but couldn't be fetched because it wasn't
committed.

We now defer the last progress notification until the commit
completes.